MP_backend/scheduler_backend/requirements.txt
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.7
django-crispy-forms==1.8.0
django-heroku==0.3.1
django-mysql==3.2.0
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
gunicorn==20.0.0
idna==2.8
MouseInfo==0.1.2
mysqlclient==1.4.5
Pillow==6.2.1
psycopg2==2.8.4
PyGetWindow==0.0.8
PyMsgBox==1.0.7
pyperclip==1.7.0
PyRect==0.1.4
PyScreeze==0.1.25
python-xlib==0.25
python3-xlib==0.15
PyTweening==1.0.3
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
whitenoise==4.1.4
